extracted construct_statefulset
