+++ operator/pkg/reconciliation/construct_statefulset_test.go
@@ -1,0 +1,50 @@
+// Copyright DataStax, Inc.
+// Please see the included license file for details.
+
+package reconciliation
+
+import (
+	"testing"
+
+	api "github.com/datastax/cass-operator/operator/pkg/apis/cassandra/v1beta1"
+	"github.com/stretchr/testify/assert"
+	corev1 "k8s.io/api/core/v1"
+)
+
+func Test_newStatefulSetForCassandraDatacenter(t *testing.T) {
+	type args struct {
+		rackName     string
+		dc           *api.CassandraDatacenter
+		replicaCount int
+	}
+	tests := []struct {
+		name string
+		args args
+	}{
+		{
+			name: "test nodeSelector",
+			args: args{
+				rackName:     "r1",
+				replicaCount: 1,
+				dc: &api.CassandraDatacenter{
+					Spec: api.CassandraDatacenterSpec{
+						ClusterName:  "c1",
+						NodeSelector: map[string]string{"dedicated": "cassandra"},
+						StorageConfig: api.StorageConfig{
+							CassandraDataVolumeClaimSpec: &corev1.PersistentVolumeClaimSpec{},
+						},
+						ServerType:    "cassandra",
+						ServerVersion: "3.11.7",
+					},
+				},
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Log(tt.name)
+		got, err := newStatefulSetForCassandraDatacenter(tt.args.rackName, tt.args.dc, tt.args.replicaCount)
+		assert.NoError(t, err, "newStatefulSetForCassandraDatacenter should not have errored")
+		assert.NotNil(t, got, "newStatefulSetForCassandraDatacenter should not have returned a nil statefulset")
+		assert.Equal(t, map[string]string{"dedicated": "cassandra"}, got.Spec.Template.Spec.NodeSelector)
+	}
+}
