+++ operator/pkg/reconciliation/constructor_test.go
@@ -49,44 +49,6 @@ func Test_calculateNodeAffinity(t *testing.T) {
 	})
 }
 
-func Test_newStatefulSetForCassandraDatacenter(t *testing.T) {
-	type args struct {
-		rackName     string
-		dc           *api.CassandraDatacenter
-		replicaCount int
-	}
-	tests := []struct {
-		name string
-		args args
-	}{
-		{
-			name: "test nodeSelector",
-			args: args{
-				rackName:     "r1",
-				replicaCount: 1,
-				dc: &api.CassandraDatacenter{
-					Spec: api.CassandraDatacenterSpec{
-						ClusterName:  "c1",
-						NodeSelector: map[string]string{"dedicated": "cassandra"},
-						StorageConfig: api.StorageConfig{
-							CassandraDataVolumeClaimSpec: &corev1.PersistentVolumeClaimSpec{},
-						},
-						ServerType:    "cassandra",
-						ServerVersion: "3.11.7",
-					},
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Log(tt.name)
-		got, err := newStatefulSetForCassandraDatacenter(tt.args.rackName, tt.args.dc, tt.args.replicaCount)
-		assert.NoError(t, err, "newStatefulSetForCassandraDatacenter should not have errored")
-		assert.NotNil(t, got, "newStatefulSetForCassandraDatacenter should not have returned a nil statefulset")
-		assert.Equal(t, map[string]string{"dedicated": "cassandra"}, got.Spec.Template.Spec.NodeSelector)
-	}
-}
-
 func TestCassandraDatacenter_buildInitContainer_resources_set(t *testing.T) {
 	dc := &api.CassandraDatacenter{
 		Spec: api.CassandraDatacenterSpec{
